+++ site/pages/our-work/index.js
@@ -2,16 +2,21 @@ import React from 'react';
 
 import styles from './style.css';
 
-import CaseStudyBankSlice from './case-study-bank-slice';
-import CaseStudyCarTrawlerSlice from './case-study-car-trawler-slice';
-import CaseStudyRetailerSlice from './case-study-retailer-slice';
-import CaseStudyCamdenSlice from './case-study-camden-slice';
-import CaseStudyFtSlice from './case-study-ft-slice';
-import CaseStudyFortnumSlice from './case-study-fortnum-and-mason-slice';
-import CaseStudyFortnumDigitalTransformationSlice from './case-study-fortnum-and-mason-digital-transformation-slice';
-import CaseStudySkySlice from './case-study-sky-slice';
-import ClientLogosSlice from './client-logos-slice';
-import CaseStudyCell from './case-study-cell';
+import BankSlice from './slices/bank';
+import CarTrawlerSlice from './slices/car-trawler';
+import RetailerSlice from './slices/retailer';
+import CamdenSlice from './slices/camden';
+import FtSlice from './slices/ft';
+import FortnumSlice from './slices/fortnum-and-mason';
+import FortnumDigitalTransformationSlice from './slices/fortnum-and-mason-digital-transformation';
+import SkySlice from './slices/sky';
+
+import Cell from './cell';
+import Logos from './logos';
+import BmwLogo from './logos/images/bmw.png';
+import BbcLogo from './logos/images/bbc.png';
+import SkyLogo from './logos/images/sky.png';
+
 import Social from '../../components/social';
 import ScrollTracker from '../../components/scroll-tracker';
 
@@ -19,9 +24,6 @@ import metaImage from './meta-image.jpg';
 import HallerImage from './images/Hall.jpg';
 import HallerLogo from './images/haller.png';
 import BmwImage from './images/BM.jpg';
-import BmwLogo from './client-logos-slice/images/bmw.png';
-import BbcLogo from './client-logos-slice/images/bbc.png';
-import SkyLogo from './client-logos-slice/images/sky.png';
 
 export default function CaseStudies() {
   const social = {
@@ -44,25 +46,25 @@ export default function CaseStudies() {
       </div>
       <div className={styles.caseStudyTopSection}>
         <div className={styles.topSlicesContainer}>
-          <CaseStudyCarTrawlerSlice />
-          <CaseStudyBankSlice />
-          <CaseStudyRetailerSlice />
-          <CaseStudyCamdenSlice />
-          <CaseStudyFtSlice />
-          <CaseStudyFortnumSlice />
-          <CaseStudyFortnumDigitalTransformationSlice />
-          <CaseStudySkySlice />
-          <ClientLogosSlice />
+          <CarTrawlerSlice />
+          <BankSlice />
+          <RetailerSlice />
+          <CamdenSlice />
+          <FtSlice />
+          <FortnumSlice />
+          <FortnumDigitalTransformationSlice />
+          <SkySlice />
+          <Logos />
           <div className={styles.grid}>
             <div className={styles.gridRow}>
-              <CaseStudyCell
+              <Cell
                 clientName="BBC"
                 clientLogo={BbcLogo}
                 headerText="Delivering a better customer experience, faster"
                 descriptionText="How the rapid prototyping model helped the BBC to uncover new ways to engage its audience."
                 routeKey="bbcCaseStudy"
               />
-              <CaseStudyCell
+              <Cell
                 clientName="Sky"
                 clientLogo={SkyLogo}
                 headerText="Helping customers help themselves"
@@ -71,7 +73,7 @@ export default function CaseStudies() {
               />
             </div>
             <div className={styles.gridRow}>
-              <CaseStudyCell
+              <Cell
                 clientName="Haller"
                 clientLogo={HallerLogo}
                 image={HallerImage}
@@ -79,7 +81,7 @@ export default function CaseStudies() {
                 descriptionText="Red Badger teamed up with the Haller Foundation on a pro-bono basis to develop a mobile application which helps Kenyan farmers."
                 routeKey="hallerCaseStudy"
               />
-              <CaseStudyCell
+              <Cell
                 clientName="BMW"
                 clientLogo={BmwLogo}
                 image={BmwImage}
