+++ site/pages/our-work/logos/images/pride.png
