+++ site/pages/our-work/logos/images/sky.png
