+++ site/pages/our-work/slices/fortnum-and-mason-digital-transformation/index.js
@@ -1,21 +1,21 @@
 import React from 'react';
 import styles from './style.css';
-import Link from '../../../components/link';
+import Link from '../../../../components/link';
 
-import skyLogo from './images/sky.png';
+import fmLogo from './images/fortnum-mason-logo.png';
 
-export default function CaseStudySkySlice() {
+export default function CaseStudyFortnumDigitalTransformationSlice() {
   return (
     <div className={styles.caseStudyContainer}>
       <div className={styles.caseStudyContent}>
         <div className={styles.caseStudyTextContainer}>
-          <Link to="skyCMSCaseStudy">
-            <img src={skyLogo} className={styles.clientLogo} alt="Sky logo" />
+          <Link to="fMTeaCaseStudy">
+            <img src={fmLogo} className={styles.clientLogo} alt="Fortnum and Mason logo" />
             <h2 className={styles.caseStudyTextContainerHeader}>
-              A modern CMS that is the foundation that supports both customers and the internal team
+              The ongoing digital transformation of a 310-year old retailer
             </h2>
             <p className={styles.description}>
-              Enabling Sky to manage their content with an amazing customer experience
+              Discover how we helped customers find products faster and drive sales of tea.
             </p>
             <div className={styles.links}>
               <p className={styles.readmore}>Read more</p>
