+++ site/pages/our-work/slices/fortnum-and-mason-digital-transformation/style.css
@@ -1,5 +1,5 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../css/_colors.css";
-@value mediumScreen, largeScreen from "../../../css/_sizes.css";
+@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
+@value mediumScreen, largeScreen from "../../../../css/_sizes.css";
 
 .caseStudyContainer {
   max-width: 1100px;
@@ -25,21 +25,21 @@
 }
 
 .caseStudyTextContainerHeader {
-  composes: fontL from "../../../css/typography/_fonts.css";
-  composes: serif from "../../../css/typography/_fonts.css";
+  composes: fontL from "../../../../css/typography/_fonts.css";
+  composes: serif from "../../../../css/typography/_fonts.css";
   color: badgerBlack;
   margin-top: 30px;
 }
 
 .description {
-  composes: fontXS from "../../../css/typography/_fonts.css";
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
+  composes: fontXS from "../../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
   color: badgerBlack;
   margin-top: 10px;
 }
 
 .readmore {
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
   border-bottom: 2px solid badgerRedOnWhite;
   text-decoration: none;
   color: badgerBlack;
