+++ site/pages/our-work/slices/ft/images/meeting.png
