+++ site/pages/our-work/slices/ft/style.css
@@ -1,5 +1,5 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../css/_colors.css";
-@value mediumScreen from "../../../css/_sizes.css";
+@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
+@value mediumScreen from "../../../../css/_sizes.css";
 
 .caseStudyContainer {
   max-width: 1100px;
@@ -43,21 +43,21 @@ _:-ms-fullscreen, :root .imageWrapper {
 }
 
 .caseStudyTextContainerHeader {
-  composes: fontM from "../../../css/typography/_fonts.css";
-  composes: serif from "../../../css/typography/_fonts.css";
+  composes: fontM from "../../../../css/typography/_fonts.css";
+  composes: serif from "../../../../css/typography/_fonts.css";
   color: badgerBlack;
   margin-top: 20px;
 }
 
 .description {
-  composes: fontXS from "../../../css/typography/_fonts.css";
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
+  composes: fontXS from "../../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
   color: badgerBlack;
   margin-top: 10px;
 }
 
 .readmore {
-  composes: boldSansSerif from "../../../css/typography/_fonts.css";
+  composes: boldSansSerif from "../../../../css/typography/_fonts.css";
   border-bottom: 2px solid badgerRedOnWhite;
   text-decoration: none;
   color: badgerBlack;
