+++ site/pages/our-work/slices/retailer/style.css
@@ -1,5 +1,5 @@
-@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../css/_colors.css";
-@value mediumScreen, largeScreen from "../../../css/_sizes.css";
+@value badgerBlack, badgerWhite, badgerRedOnWhite, linesOnWhite from "../../../../css/_colors.css";
+@value mediumScreen, largeScreen from "../../../../css/_sizes.css";
 
 .imageWrapper {
   display: flex;
@@ -32,21 +32,21 @@ _:-ms-fullscreen,
 }
 
 .caseStudyTextContainerHeader {
-  composes: fontM from '../../../css/typography/_fonts.css';
-  composes: serif from '../../../css/typography/_fonts.css';
+  composes: fontM from '../../../../css/typography/_fonts.css';
+  composes: serif from '../../../../css/typography/_fonts.css';
   color: badgerBlack;
   margin-top: 20px;
 }
 
 .description {
-  composes: fontXS from '../../../css/typography/_fonts.css';
-  composes: boldSansSerif from '../../../css/typography/_fonts.css';
+  composes: fontXS from '../../../../css/typography/_fonts.css';
+  composes: boldSansSerif from '../../../../css/typography/_fonts.css';
   color: badgerBlack;
   margin-top: 10px;
 }
 
 .readmore {
-  composes: boldSansSerif from '../../../css/typography/_fonts.css';
+  composes: boldSansSerif from '../../../../css/typography/_fonts.css';
   border-bottom: 2px solid badgerRedOnWhite;
   text-decoration: none;
   color: badgerBlack;
@@ -78,14 +78,14 @@ _:-ms-fullscreen,
 }
 
 .imagePercentage {
-  composes: fontM from '../../../css/typography/_fonts.css';
-  composes: serif from '../../../css/typography/_fonts.css';
+  composes: fontM from '../../../../css/typography/_fonts.css';
+  composes: serif from '../../../../css/typography/_fonts.css';
   color: badgerBlack;
 }
 
 .imageText {
-  composes: fontS from '../../../css/typography/_fonts.css';
-  composes: lightSerif from '../../../css/typography/_fonts.css';
+  composes: fontS from '../../../../css/typography/_fonts.css';
+  composes: lightSerif from '../../../../css/typography/_fonts.css';
   color: badgerBlack;
   font-size: 16px;
 }
