+++ site/pages/our-work/slices/sky/images/sky.png
